+++ lib/utils/format.js
