+++ lib/utils/sum_edits.js
