+++ lib/utils/trim_length.js
