+++ pages/about.js
@@ -4,9 +4,7 @@ import Link from 'next/link';
 const image1 = '/static/about-campaignmanager.jpg';
 const image2 = '/static/about-mapper.jpg';
 
-const projectName = process.env.PROJECT_NAME || 'OpenStreetMap'
-
-const About = () => (
+const About = ({project}) => (
   <div className="About">
     <header className="header--internal ">
       <div className="row">
@@ -15,7 +13,7 @@ const About = () => (
     </header>
     <section className="text-body section-first--sm">
       <div className="row">
-        <p className="text-body--large">{'Scoreboard is a site that provides insight into the ' + projectName + ' ecosystem, making the users not only more informed but more efficient.'}</p>
+        <p className="text-body--large">{'Scoreboard is a site that provides insight into the ' + project + ' ecosystem, making the users not only more informed but more efficient.'}</p>
         <div className="clearfix about--block">
           <p style={{float: "left"}} className="text-body--med width--sm">For campaign managers, Scoreboard provides an overview of what’s happening on the system, and also a way to get detailed information on individual mappers or campaigns. Discover all active <Link href="/campaigns"><a>campaigns.</a></Link></p>
           <Link href="/campaigns">
